+++ src/account/mod.rs
@@ -1,3 +1,5 @@
 pub mod account;
+mod chat;
 mod chatlistitem;
 mod message;
+mod util;
